Add lookupSourcePositions to JoinOperatorInfo
Also add implementation of getJoinPositionCount to
PartitionedLookupSource. PartitionedLookupSource can have an array of
lookup sources, each of which return an integer for
getJoinPositionCount an must be summed together, so this also required
changing the return type for the interface to a long.

+++ presto-main/src/main/java/com/facebook/presto/operator/JoinHash.java
@@ -55,7 +55,7 @@ public final class JoinHash
     }
 
     @Override
-    public int getJoinPositionCount()
+    public long getJoinPositionCount()
     {
         return pagesHash.getPositionCount();
     }
