+++ presto-main/src/main/java/com/facebook/presto/operator/index/IndexLoader.java
@@ -397,7 +397,7 @@ public class IndexLoader
         }
 
         @Override
-        public int getJoinPositionCount()
+        public long getJoinPositionCount()
         {
             return 0;
         }
